+++ src/frontend/apps/standalone_site/src/features/Contents/features/ClassRoom/components/Read/ClassRooms.tsx
@@ -1,9 +1,13 @@
 import { useClassrooms } from 'lib-classroom';
-import { useState } from 'react';
+import { Fragment, useState } from 'react';
 import { defineMessages, useIntl } from 'react-intl';
 
 import { REACT_QUERY_CONF_API } from 'conf/global';
-import { ContentsWrapper, useContentPerPage } from 'features/Contents/';
+import {
+  ContentsFilter,
+  ContentsWrapper,
+  useContentPerPage,
+} from 'features/Contents/';
 
 import ClassRoomItem from './ClassRoomItem';
 
@@ -24,29 +28,38 @@ const ClassRooms = ({ withPagination = true, limit }: ClassRoomsProps) => {
   const intl = useIntl();
   const [currentPage, setCurrentPage] = useState(1);
   const contentPerPage = useContentPerPage();
+  const [filter, setFilter] = useState<{
+    playlist: string;
+  }>({
+    playlist: '',
+  });
 
   const apiResponse = useClassrooms(
     {
       offset: `${(currentPage - 1) * contentPerPage}`,
       limit: `${limit || contentPerPage}`,
+      playlist: filter.playlist,
     },
     REACT_QUERY_CONF_API,
   );
 
   return (
-    <ContentsWrapper
-      apiResponse={apiResponse}
-      dataComponent={(classroom, index) => (
-        <ClassRoomItem
-          key={`classroom-${classroom.id}-${index}`}
-          classroom={classroom}
-        />
-      )}
-      currentPage={currentPage}
-      setCurrentPage={(page) => setCurrentPage(page)}
-      noContentMessage={intl.formatMessage(messages.NoClassroom)}
-      withPagination={withPagination}
-    />
+    <Fragment>
+      <ContentsFilter setFilter={(newFilter) => setFilter(newFilter)} />
+      <ContentsWrapper
+        apiResponse={apiResponse}
+        dataComponent={(classroom, index) => (
+          <ClassRoomItem
+            key={`classroom-${classroom.id}-${index}`}
+            classroom={classroom}
+          />
+        )}
+        currentPage={currentPage}
+        setCurrentPage={(page) => setCurrentPage(page)}
+        noContentMessage={intl.formatMessage(messages.NoClassroom)}
+        withPagination={withPagination}
+      />
+    </Fragment>
   );
 };